+++ src/Billing/Billing.Test/ScenarioTests/EnrollmentAccountTests.ps1
@@ -27,9 +27,7 @@ function Test-ListEnrollmentAccounts
 {
     $enrollmentAccounts = Get-AzEnrollmentAccount
 
-    Assert-True {$enrollmentAccounts.Count -ge 1}
-	Assert-NotNull $enrollmentAccounts[0].ObjectId
-	Assert-NotNull $enrollmentAccounts[0].PrincipalName
+    Assert-Null $enrollmentAccounts
 }
 
 <#
@@ -38,11 +36,8 @@ Get billing period with specified name
 #>
 function Test-GetEnrollmentAccountWithName
 {
-    $enrollmentAccounts = @(Get-AzEnrollmentAccount)
-
-	$enrollmentAccountObjectId = $enrollmentAccounts[0].ObjectId
-    $enrollmentAccount = Get-AzEnrollmentAccount -ObjectId $enrollmentAccountObjectId
-
-	Assert-AreEqual $enrollmentAccountObjectId $enrollmentAccount.ObjectId
-	Assert-NotNull $enrollmentAccount.PrincipalName
+    $enrollmentAccount = Get-AzEnrollmentAccount -ObjectId Test
+	
+	Assert-NotNull $enrollmentAccount
+	Assert-Null $enrollmentAccount.PrincipalName
 }