+++ src/Billing/Billing.Test/ScenarioTests/InvoicesTests.ps1
@@ -12,6 +12,8 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------------
 
+
+# Legacy invoicing tests
 <#
 .SYNOPSIS
 List invoices
@@ -20,14 +22,12 @@ function Test-ListInvoices
 {
     $billingInvoices = Get-AzBillingInvoice
 
-    Assert-True {$billingInvoices.Count -ge 1}
-	Assert-NotNull $billingInvoices[0].Name
-	Assert-NotNull $billingInvoices[0].Id
-	Assert-NotNull $billingInvoices[0].Type
-	Assert-NotNull $billingInvoices[0].InvoicePeriodStartDate
-	Assert-NotNull $billingInvoices[0].InvoicePeriodEndDate
-	Assert-Null $billingInvoices[0].DownloadUrl
-	Assert-Null $billingInvoices[0].DownloadUrlExpiry
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.BillingProfileId
 }
 
 <#
@@ -38,14 +38,12 @@ function Test-ListInvoicesWithDownloadUrl
 {
     $billingInvoices = Get-AzBillingInvoice -GenerateDownloadUrl
 
-    Assert-True {$billingInvoices.Count -ge 1}
-	Assert-NotNull $billingInvoices[0].Name
-	Assert-NotNull $billingInvoices[0].Id
-	Assert-NotNull $billingInvoices[0].Type
-	Assert-NotNull $billingInvoices[0].InvoicePeriodStartDate
-	Assert-NotNull $billingInvoices[0].InvoicePeriodEndDate
-	Assert-NotNull $billingInvoices[0].DownloadUrl
-	Assert-NotNull $billingInvoices[0].DownloadUrlExpiry
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.DownloadUrl
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
 }
 
 <#
@@ -56,14 +54,13 @@ function Test-ListInvoicesWithMaxCount
 {
     $billingInvoices = Get-AzBillingInvoice -GenerateDownloadUrl -MaxCount 1
 
+    Assert-NotNull $billingInvoices
     Assert-True {$billingInvoices.Count -eq 1}
-	Assert-NotNull $billingInvoices[0].Name
-	Assert-NotNull $billingInvoices[0].Id
-	Assert-NotNull $billingInvoices[0].Type
-	Assert-NotNull $billingInvoices[0].InvoicePeriodStartDate
-	Assert-NotNull $billingInvoices[0].InvoicePeriodEndDate
-	Assert-NotNull $billingInvoices[0].DownloadUrl
-	Assert-NotNull $billingInvoices[0].DownloadUrlExpiry
+    Assert-NotNull $billingInvoices.DownloadUrl
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
 }
 
 <#
@@ -72,41 +69,217 @@ Get latest invoice
 #>
 function Test-GetLatestInvoice
 {
-    $invoice = Get-AzBillingInvoice -Latest
+    $billingInvoice = Get-AzBillingInvoice -Latest -GenerateDownloadUrl
 
-	Assert-NotNull $invoice.Name
-	Assert-NotNull $invoice.Id
-	Assert-NotNull $invoice.Type
-	Assert-NotNull $invoice.InvoicePeriodStartDate
-	Assert-NotNull $invoice.InvoicePeriodEndDate
-	Assert-NotNull $invoice.DownloadUrl
-	Assert-NotNull $invoice.DownloadUrlExpiry
+	Assert-NotNull $billingInvoice	
+    Assert-NotNull $billingInvoice.Name
+    Assert-NotNull $billingInvoice.Status
+    Assert-NotNull $billingInvoice.BillingProfileDisplayName
+    Assert-NotNull $billingInvoice.InvoiceDate
 }
 
 <#
 .SYNOPSIS
-Get invoice with specified name
+Get invoice with specified name with GenerateDownloadUrl
 #>
-function Test-GetInvoiceWithName
+function Test-GetInvoiceByNameWithDownloadUrl
 {
-    $sampleInvoices = Get-AzBillingInvoice
-    Assert-True { $sampleInvoices.Count -ge 1 }
+	$billingInvoices = Get-AzBillingInvoice -Name T000512627 -GenerateDownloadUrl
 
-	$invoice = Get-AzBillingInvoice -Name $sampleInvoices[0].Name
+	Assert-NotNull $billingInvoices
+	Assert-NotNull $billingInvoices.Name
+    Assert-AreEqual $billingInvoices.Name T000512627
+}
 
-	Assert-AreEqual $invoice.Id $sampleInvoices[0].Id
+# Modern Invoicing Tests
+<#
+.SYNOPSIS
+Get modern invoice by InvoiceName
+#>
+function Test-GetModernInvoiceByName
+{
+    $billingInvoices = Get-AzBillingInvoice -Name T000512627
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.DownloadUrl
 }
 
 <#
 .SYNOPSIS
-Get invoice with specified names
+List invoices by billingAccountName
 #>
-function Test-GetInvoiceWithNames
+function Test-ListModernInvoicesByBillingAccountName
 {
-	$sampleInvoices = Get-AzBillingInvoice
-    Assert-True { $sampleInvoices.Count -gt 1 }
+    $billingInvoices = Get-AzBillingInvoice -BillingAccountName db038d21-b0d2-463c-942f-b09127c6f4e4:7c9c4a38-593e-479e-8958-9a338a0d8d02_2019-05-31
 
-    $billingInvoices = Get-AzBillingInvoice -Name $sampleInvoices.Name
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.DownloadUrl
+}
 
-    Assert-AreEqual $sampleInvoices.Count $billingInvoices.Count
+<#
+.SYNOPSIS
+List invoices by billingAccountName with DownloadUrl
+#>
+function Test-ListModernInvoicesByBillingAccountNameWithDownloadUrl
+{
+    $billingInvoices = Get-AzBillingInvoice -BillingAccountName db038d21-b0d2-463c-942f-b09127c6f4e4:7c9c4a38-593e-479e-8958-9a338a0d8d02_2019-05-31 -GenerateDownloadUrl
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-NotNull $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+List invoices by billingAccountName
+#>
+function Test-ListModernInvoicesByBillingAccountNameWithMaxCount
+{
+    $billingInvoices = Get-AzBillingInvoice -MaxCount 1 -BillingAccountName db038d21-b0d2-463c-942f-b09127c6f4e4:7c9c4a38-593e-479e-8958-9a338a0d8d02_2019-05-31 
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+Get latest invoice by BillingAccountName
+#>
+function Test-GetLatestModernInvoiceByBillingAccountName
+{
+    $billingInvoices = Get-AzBillingInvoice -Latest -BillingAccountName db038d21-b0d2-463c-942f-b09127c6f4e4:7c9c4a38-593e-479e-8958-9a338a0d8d02_2019-05-31
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+Get latest invoice by BillingAccountName with DownloadUrl
+#>
+function Test-GetLatestModernInvoiceByBillingAccountNameWithDownloadUrl
+{
+    $billingInvoices = Get-AzBillingInvoice -Latest -BillingAccountName db038d21-b0d2-463c-942f-b09127c6f4e4:7c9c4a38-593e-479e-8958-9a338a0d8d02_2019-05-31 -GenerateDownloadUrl
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-NotNull $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+Get latest invoice by BillingAccountName and InvoiceName with DownloadUrl
+#>
+function Test-GetModernInvoiceByBillingAccountNameAndInvoiceNameWithDownloadUrl
+{
+    $billingInvoices = Get-AzBillingInvoice -Name T000512627 -BillingAccountName db038d21-b0d2-463c-942f-b09127c6f4e4:7c9c4a38-593e-479e-8958-9a338a0d8d02_2019-05-31 -GenerateDownloadUrl
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-NotNull $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+Get latest invoice by BillingAccountName and InvoiceName and do not generate DownloadUrl
+#>
+function Test-GetModernInvoiceByBillingAccountNameAndInvoiceName
+{
+    $billingInvoices = Get-AzBillingInvoice -Name T000512627 -BillingAccountName db038d21-b0d2-463c-942f-b09127c6f4e4:7c9c4a38-593e-479e-8958-9a338a0d8d02_2019-05-31 
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+List invoices by billingProfileName
+#>
+function Test-ListModernInvoicesByBillingProfileName
+{
+    $billingInvoices = Get-AzBillingInvoice -BillingAccountName c017063b-18ad-5e26-f4af-a4d7eff204cb:171df24e-c924-4c58-9daa-a0bdb1686fef_2019-05-31 -BillingProfileName PO6F-IWMU-BG7-TGB
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+List invoices by billingProfileName with DownloadUrl
+#>
+function Test-ListModernInvoicesByBillingProfileNameWithDownloadUrl
+{
+    $billingInvoices = Get-AzBillingInvoice -BillingAccountName c017063b-18ad-5e26-f4af-a4d7eff204cb:171df24e-c924-4c58-9daa-a0bdb1686fef_2019-05-31 -BillingProfileName PO6F-IWMU-BG7-TGB -GenerateDownloadUrl
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-NotNull $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+List invoices by billingProfileName with MaxCount
+#>
+function Test-ListModernInvoicesByBillingProfileNameMaxCount
+{
+    $billingInvoices = Get-AzBillingInvoice -BillingAccountName c017063b-18ad-5e26-f4af-a4d7eff204cb:171df24e-c924-4c58-9daa-a0bdb1686fef_2019-05-31 -BillingProfileName PO6F-IWMU-BG7-TGB -MaxCount 1
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.DownloadUrl
+}
+
+<#
+.SYNOPSIS
+Get latest invoice by BillingProfileName
+#>
+function Test-GetLatestInvoicesByBillingProfileName
+{
+    $billingInvoices = Get-AzBillingInvoice -Latest -BillingAccountName c017063b-18ad-5e26-f4af-a4d7eff204cb:171df24e-c924-4c58-9daa-a0bdb1686fef_2019-05-31 -BillingProfileName PO6F-IWMU-BG7-TGB
+
+    Assert-NotNull $billingInvoices
+    Assert-NotNull $billingInvoices.Name
+    Assert-NotNull $billingInvoices.Status
+    Assert-NotNull $billingInvoices.BillingProfileDisplayName
+    Assert-NotNull $billingInvoices.InvoiceDate
+    Assert-Null $billingInvoices.DownloadUrl
 }