+++ src/Billing/Billing/Models/PSBillingPeriod.cs
@@ -16,7 +16,6 @@ using Microsoft.Azure.Commands.Billing.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using ApiBillingPeriod = Microsoft.Azure.Management.Billing.Models.BillingPeriod;
 
 namespace Microsoft.Azure.Commands.Billing.Models
 {
@@ -36,22 +35,6 @@ namespace Microsoft.Azure.Commands.Billing.Models
 
         public PSBillingPeriod()
         {
-        }
-
-        public PSBillingPeriod(ApiBillingPeriod billingPeriod)
-        {
-            if (billingPeriod != null)
-            {
-                this.Id = billingPeriod.Id;
-                this.Type = billingPeriod.Type;
-                this.Name = billingPeriod.Name;
-                this.BillingPeriodStartDate = billingPeriod.BillingPeriodStartDate;
-                this.BillingPeriodEndDate = billingPeriod.BillingPeriodEndDate;
-                if (billingPeriod.InvoiceIds != null)
-                {
-                    this.InvoiceNames = billingPeriod.InvoiceIds.Select(x => Utilities.GetResourceNameFromId(x)).ToList();
-                }
-            }
         }
     }
 }